logout and fixing post deletion

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -35,6 +36,9 @@ class PostController extends Controller
   public function getDeletePost($post_id)
   {
     $post = Post::where('id', $post_id)->first();
+    if (Auth::user() != $post->user) {
+      return redirect()->back();
+    }
     $post->delete();
     return redirect()->route('dashboard')->with(['message' => "Successfully deleted"]);
   }
